src/main/java/com/example/CapstoneProject/Service/UserService.java
package com.example.CapstoneProject.Service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.example.CapstoneProject.Repository.UserRepository;
import com.example.CapstoneProject.entity.User;

@Service
public class UserService {
    @Autowired
    private UserRepository userRepository;

    public User registerUser(User user) {
        // Add logic for user registration (e.g., save user to database)
        return userRepository.save(user);
    }

    public User findByUsername(String username) {
        return userRepository.findByUsername(username);
    }
}
